Repair build-all.sh and clean-all.sh.
Make it so that it runs ../configure and make for each target. This makes the wrapper-Makefile obsolete.
It moves the binaries to tools/checkwps/output. NOTE: Something is fishy with the creative players (it builds them, but can't move them), because the configure target name and modelname differs (not sure how to fix that).

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@22696 a1c6a512-1295-4272-9138-f99709370657