WPS: Revert some of r24193 as decided in FS#10882.
This commit reverts having the track progress bar mirrored horizontally. This
is done because the code that makes the progress bar sensitive to language
direction conflicts with direction of buttons used to control playback, i.e.
progress bar fills to the left, but pressing long left makes it move to the
right, towards the beginning of the track.
The code that swap the progress bar direction was not (and will not be)
committed.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24196 a1c6a512-1295-4272-9138-f99709370657