Simulator: Make touchscreen targets usable on keyboards w/o numpad (laptops)
Use 7,8,9,u,i,o,j,k,l as equivalent to the numpad keys used to simulate
touching in grid mode (3x3 buttons). These are the keys usually used on such
keyboards for numpad keys when the Num Lock is enabled.

The USB key on these targets is changed to 'c' (connect) to avoid conflict in
key mappings.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24252 a1c6a512-1295-4272-9138-f99709370657