Enable charging monitoring for the Nano 1G.
The code for this was already present as it seems to behave the same as the Video, but CONFIG_CHARGING had not been set, probably for lack of testing. It has now been tested.

This allows the WPS/status bar to correctly display the charging status, and also fixes various bugs:
FS#6891 iPod Nano Charger Support
FS#6940 Nano Car Adapter Mode enabled
FS#10591 iPod nano reboots on idle poweroff when plugged in


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24581 a1c6a512-1295-4272-9138-f99709370657