Prevent accidental skipping to beginning/end of file in text viewer.
When the text viewer is not in wide mode, the buttons mapped to move the screen left and right instead jump to the top and bottom of the file. This is very easy to do by accident on most targets, causing you to lose your place. This change requires that the button be held for a long press before it jumps to the top/bottom; short presses simply trigger paging up and down by one page. There is no change in behaviour when wide mode is enabled.

Based on an idea from FS#2648 but implemented differently.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@25262 a1c6a512-1295-4272-9138-f99709370657