Warn when selecting system proxy settings with invalid values.
Rockbox Utility doesn't support "proxy auto-config" (aka PAC) for system proxy
settings. This can result in users selecting system proxy and expecting it to
work even if their system uses PAC. While the configuration dialog displays the
proxy setting values retrieved from the system this is not totally obvious. Add
a message telling if the retrieved system proxy values are not useable. For now
this only checks if a proxy host and port are set which should catch the usual
cases when PAC is used.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@25435 a1c6a512-1295-4272-9138-f99709370657