Add repeat counting to system trace.
This make the system trace check for repeating lines, and only print a "(last
message repeated n lines.)" instead of the real message multiple times. This
keeps the trace much shorter if messages are repeated. The drawback is that the
replacement count message will only get printed on the next line getting
traced, so until that happens it swallows the repeated lines.
Before saving the systrace buffer is flushed, so this should not raise a problem.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@25878 a1c6a512-1295-4272-9138-f99709370657