Support resolving of DLLs when running on Windows.
Resolve the DLLs required by the built executable and try to add the required
DLL files that are not recognized as system libraries to the resulting zip /
NSIS installer. This means that it's now possible to easily build both Theme
Editor and Rockbox Utility as dynamically linked binary without the risk of
missing required DLLs in the package. The major advantage of this is that
it's not necessary anymore to have a statically built Qt installation for
building releases. The drawback is that the created binaries will rely on
additional DLL files, so it's no longer a single-run binary.

Binary release of Rockbox Utility should still be statically build.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@27945 a1c6a512-1295-4272-9138-f99709370657