debian directory needed to build maemo packages
The build rules in the debain directoy are currently
maemo specific and won't build on debian.

Later on we could add automatic distribution detection
and switch the build rules based on that if someone
wants to build a generic RaaA package for debian.

Maemo binary packages get build like this (in scratchbox):

dpkg-buildpackage -rfakeroot -sa

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@29256 a1c6a512-1295-4272-9138-f99709370657