Better dircache handling in simulator
Actually add files to the dircache (as well as removing them) and
remove them only when they were found on disk.  This matches the
native behavior and prevents the dircache from becoming uninitialized
when a previously created or nonexisting file is being removed.  This
typically would happen during a database update.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@29715 a1c6a512-1295-4272-9138-f99709370657