Manual: add search box to HTML output.
To allow easy searching the HTML (multi page) output add a search box that
queries Google. The Google search will get restricted to the location the
manual is stored. Google seems to happily accept a subfolder when specifying a
domain to search. For this to work the search box has to be created dynamically
with the help of a bit JavaScript.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@30017 a1c6a512-1295-4272-9138-f99709370657