(experiment) don't shrink in buflib for Android & reduce the Audiobuffer size
It is still uncertain why the memory size would alter the sound signiture.
my test result shows that, by lowering the ram to 16mb can relief this symptom. kugel's dynamic memory allocation patch also shows not shrink helps.

may need to take this with a grain of salt and may needs to revered this patch in the future.